make nvtx an implementation of profile
