+++ verl/utils/debug/profile.py
@@ -96,6 +96,21 @@ class Profiler:
             self.skip_prof = True
 
 
+def mark_start_range(message: Optional[str] = None, color: Optional[str] = None, domain: Optional[str] = None, category: Optional[str] = None) -> None:
+    pass
+
+
+def mark_end_range(range_id: str) -> None:
+    pass
+
+
+def mark_annotate(message: Optional[str] = None, color: Optional[str] = None, domain: Optional[str] = None, category: Optional[str] = None) -> Callable:
+    def decorator(func):
+        return func
+
+    return decorator
+
+
 @dataclass
 class ProfilerConfig:
     """
@@ -121,71 +136,22 @@ class ProfilerConfig:
         )
 
 
-class NsightSystemsProfiler:
-    """
-    Nsight system profiler. Installed in a worker to control the Nsight system profiler.
-    """
-
+class WorkerProfiler:
     def __init__(self, rank: int, config: Optional[ProfilerConfig] = None):
-        config = config or ProfilerConfig()
-        self.this_step: bool = False
-        self.discrete: bool = config.discrete
-        self.this_rank: bool = rank in (config.ranks or []) or config.all_ranks
+        pass
 
     def start(self):
-        if self.this_rank:
-            self.this_step = True
-            if not self.discrete:
-                torch.cuda.profiler.start()
+        pass
 
     def stop(self):
-        if self.this_rank:
-            self.this_step = False
-            if not self.discrete:
-                torch.cuda.profiler.stop()
+        pass
 
     @staticmethod
     def annotate(message: Optional[str] = None, color: Optional[str] = None, domain: Optional[str] = None, category: Optional[str] = None) -> Callable:
-        """Decorate a Worker member function to profile the current rank in the current training step.
-
-        Requires the target function to be a member function of a Worker, which has a member field `profiler` with NightSystemsProfiler type.
-
-        Args:
-            message (str, optional):
-                The message to be displayed in the profiler. Defaults to None.
-            color (str, optional):
-                The color of the range. Defaults to None.
-            domain (str, optional):
-                The domain of the range. Defaults to None.
-            category (str, optional):
-                The category of the range. Defaults to None.
-        """
-
         def decorator(func):
             @functools.wraps(func)
             def wrapper(self, *args, **kwargs):
-                from verl.utils.import_utils import is_nvtx_available
-
-                if is_nvtx_available():
-                    from .nvtx_annotations import mark_end_range, mark_start_range
-                else:
-                    from .empty_annotations import mark_end_range, mark_start_range
-
-                profile_name = message or func.__name__
-
-                if self.profiler.this_step:
-                    if self.profiler.discrete:
-                        torch.cuda.profiler.start()
-                    mark_range = mark_start_range(message=profile_name, color=color, domain=domain, category=category)
-
-                result = func(self, *args, **kwargs)
-
-                if self.profiler.this_step:
-                    mark_end_range(mark_range)
-                    if self.profiler.discrete:
-                        torch.cuda.profiler.stop()
-
-                return result
+                return func(self, *args, **kwargs)
 
             return wrapper
 
@@ -193,7 +159,7 @@ class NsightSystemsProfiler:
 
 
 class WorkerProfilerExtension:
-    def __init__(self, profiler: NsightSystemsProfiler):
+    def __init__(self, profiler: WorkerProfiler):
         self.profiler = profiler
 
     from verl.single_controller.base.decorator import Dispatch, register
